requirements.txt
flask==3.1.1
langchain==0.3.26
langchain-core==0.3.80
langchain-community==0.3.26
langchain-text-splitters==0.3.11
langchain-google-genai==2.0.10
langchain-pinecone==0.2.8
langchain-groq==0.2.0
langsmith==0.4.49
pinecone[asyncio]==7.3.0
pypdf==5.6.1
python-dotenv==1.1.0
sentence-transformers==2.3.1
torch==2.2.0
transformers==4.36.2
